nss-qdgym/requirements.txt
# mujoco_py is not included here, but we install a patched fork of it from
# https://github.com/icaros-usc/mujoco-py

ribs[all]==0.4.0

gym==0.18.0
Box2D==2.3.10
pybullet==3.1.7
numpy==1.20.2
scipy==1.7.2
pandas==1.2.3
numba==0.52.0
gin-config==0.4.0
fire==0.4.0
scikit-learn==0.24.1
cma==3.0.3
pymoo==0.5.0
logdir==0.11.0
loguru==0.5.3
matplotlib==3.5.1
# TODO: Update to 0.12.0 when it comes out and remove development version.
seaborn==0.11.1
alive-progress==1.6.2
threadpoolctl==2.1.0
cloudpickle==1.6.0
ruamel.yaml == 0.17.7
torch==1.9.0
python-slugify==5.0.2
statsmodels==0.13.0
pingouin==0.4.0
Pillow==7.2.0
PyVirtualDisplay==2.2

# Debugging
pyinstrument==3.3.0
pytest==6.2.2
pytest-cov==2.10.1

# Dask
dask==2.30.0
distributed==2.30.0
click==7.1.2  # Newer click causes error with dask scheduler.
bokeh==2.2.3
jupyter-server-proxy==1.5.0
